fix: move interface FragmentStrategy to types folder
